Reduce the sleep length in rgblight_disable_noeeprom (etc.)
I have no idea of why this sleep is needed, but reducing it from
50ms to 1ms seems not to have had any adverse effect for me.

The original 50ms sleep period was the cause for my split STM32F411 board
running into situation where the slave would appear deadlocked after a
rgblight_toggle().

Such a long sleep, longer than the SERIAL_USART_TIMEOUT (20ms), would
cause the master to time out and abort any transaction that had
started in this period.

It would be aborted after transmitting the transaction_id, but the
slave would be oblivious to this and respond anyway. Eventually the
master and the slave would get unsynchronized and the client's
messages would be paired up with the wrong requests.